% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeViews.R
\name{makeViews}
\alias{makeViews}
\title{makeViews}
\usage{
makeViews(
  new_env = T,
  export_views = F,
  export_path = NA,
  zip_name = "FFI_view_export"
)
}
\arguments{
\item{new_env}{Logical. If TRUE (default), will assign views into an environment called FFI_views.
If FALSE, will assign views to global environment.}

\item{export_views}{Logical. If TRUE, will export a zip file of csvs to specified export_path for
the flattened views of the data. Default is FALSE.}

\item{export_path}{Quoted string to export zipped csvs to if export = TRUE. If not specified,
will export to the working directory.}

\item{zip_name}{Quoted string that names the exported zip file (eg. "FFI_RA_AGFO_views").
If not specified, default name will be "FFI_view_export_YYYYMMDD.zip", with the
YYYYMMMDD being a date stamp for when the data were exported. Best practice would
be to name it the same as the dbname used for the importData() step.}
}
\value{
Returns flattened views of FFI data by protocol
}
\description{
Makes flattened datasets (i.e. views) that join together macroplot, sample event,
sample data and sample attribute data for Cover Points (metric), Cover Species Composition,
Density Belts (metric), Density Quadrats (metric), Disturbance History, Surface Fuels 1000Hr, Fine, and Duff,
and Trees (metric). The intent of these views is for them to be stand alone, analysis ready
datasets. Must import FFI data tables first via importData() function. Note that this has only been
tested for a few parks, and may need to be tweaked to accommodate other FFI datasets.
}
\examples{
\dontrun{
#++++ RUN FIRST ++++
library(utilsFFI)
importData(dbname = "FFI_RA_BADL")

# generate views for BADL data using defaults
makeViews()

# generate views and export zip to working directory
makeViews(export_views = T)

# generate views and export zip to specified path and name
makeViews(export_views = T, export_path = "C:/temp", zip_name = "FFI_BADL_view_export")

}


}
