% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportData.R
\name{exportData}
\alias{exportData}
\title{exportData}
\usage{
exportData(zip_name = "FFI_table_export", path = getwd())
}
\arguments{
\item{zip_name}{Quoted string that names the exported zip file (eg. "FFI_RA_AGFO").
If not specified, default name will be "FFI_table_export_YYYYMMDD.zip", with the
YYYYMMMDD being a date stamp for when the data were exported. Best practice would
be to name it the same as the dbname used for the importData() step.}

\item{path}{Quoted string to export zipped csvs to if export = TRUE.
If not specified, will export to the working directory.}
}
\value{
Saves a zip file of database tables in an imported FFI database
}
\description{
Exports data tables that were imported from the FFI database
tables in the 'dbo' schema as a zip file of csvs. Currently only works with one
park/project at a time, but could be extended to import multiple parks/projects
(see importData function in katemmiller/vegcomNGPN).
}
\examples{
\dontrun{
#--- From Local install of FFI SQL databases
# Import data for AGFO
library(utilsFFI)
importData(dbname = "FFI_RA_AGFO")

# Export data using default settings:
exportData()

# Export data to a zip file named FFI_RA_AGFO to your working directory.
exportData(zip_name = "FFI_RA_AGFO")

# Export data to a named zip file and a specific directory
exportData(zip_name = "FFI_RA_AGFO", path = "C:/temp")
}

}
