% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{importData}
\usage{
importData(type = "local", server = NA, dbname = "FFI_RA_AGFO", new_env = T)
}
\arguments{
\item{type}{Indicate how to import the database tables
\describe{
\item{"local"}{Import tables in 'dbo' schema from the local installation of an
FFI database in SQL Server Management Studio (SSMS).}
\item{"server"}{[**NOT YET ENABLED**] Import tables in 'dbo' schema from the
FFI database on the production SQL Server).}
 }}

\item{dbname}{Quoted name of database matching the name of the database as it
appears in SQL Server Management Studio
(eg. "FFI_RA_AGFO").}

\item{new_env}{Logical. If TRUE (default), will import tables to an environment
named FFI_tables. If FALSE, will import tables to global environment.}
}
\value{
Either an environment with database tables as data frames for each imported database, or database
tables directly in the global environment.
}
\description{
Imports data from FFI SQL Server database or csvs of FFI database
tables in the 'dbo' schema. Currently can only import from a local installation
of a park's FFI database in SQL Server Management Studio (SSMS), but the goal is to
add an option for importing directly from the SQL Server where the production
FFI databases are housed. Currently only works with one park/project at a time,
but could be extended to import multiple parks/projects (see importData function
in katemmiller/vegcomNGPN).
}
\examples{
\dontrun{
#--- From Local install of FFI SQL databases
# Import data for AGFO
importData(dbname = c("FFI_RA_AGFO"))

# Check that the import worked by listing names of tables in FFI_tables
names(FFI_tables)

# View data in the MacroPlot table
head(FFI_tables$MacroPlot)
str(FFI_tables$MacroPlot)

# Make the macro table in FFI_RA_AGFO a separate data frame
macro <- get("MacroPlot", envir = FFI_tables)
head(macro)
}

}
